Local Services Platform/requirements.txt
click==7.1.2
config==0.5.1
connect==0.2
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
six==1.15.0
Werkzeug==1.0.1
